Some initial norm work
